requirements.txt
gunicorn==20.0.4
absl-py==0.9.0
aiohttp==3.6.2
astor==0.8.1  
astroid==2.2.5
async-timeout==3.0.1
attrs==19.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.1
cycler==0.10.0
discord.py==1.3.2
Flask==1.1.1
gast==0.2.2
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.0
mccabe==0.6.1
multidict==4.7.5
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.2.0
pandas==1.0.1
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyHamcrest==2.0.2
pylint==2.3.1
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
selenium==3.141.0
six==1.12.0
termcolor==1.1.0
typed-ast==1.4.0
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
yarl==1.4.2
youtube-dl == 2020.3.24
pymongo == 3.10.1
setuptools == 41.6.0
dnspython == 1.16.0
mcstatus
